Add support for uPickle custom configuration
